UHF-7029: UHF-7486: Move global translations to separate file
