+++ src/js/calculator/_globalTranslations.js
@@ -1,0 +1,122 @@
+/* eslint-disable no-template-curly-in-string */
+
+const globalTranslations = {
+  has_required_fields: {
+    fi: 'Pakolliset kentät on merkitty tähdellä ${required}.',
+    sv: null,
+    en: null,
+  },
+  not_saved: {
+    fi: 'Huomioithan, että laskuriin antamiasi tietoja ei tallenneta eikä lähetetä eteenpäin.',
+    sv: null,
+    en: null,
+  },
+  unit_euro: {
+    fi: 'euroa',
+    sv: 'i euro',
+    en: 'in euros',
+  },
+  unit_person: {
+    fi: 'henkeä',
+    sv: 'i personer',
+    en: 'persons',
+  },
+  unit_day: {
+    fi: 'päivää',
+    sv: null,
+    en: null,
+  },
+  unit_hour: {
+    fi: 'tuntia',
+    sv: 'i timmar',
+    en: 'in hours',
+  },
+  unit_times_per_month: {
+    fi: 'krt / kk',
+    sv: null,
+    en: null,
+  },
+  unit_amount_per_month: {
+    fi: 'kpl / kk',
+    sv: null,
+    en: null,
+  },
+  per_item_variable: {
+    fi: 'à ${amount} ',
+    sv: null,
+    en: null,
+  },
+  per_item_unit_euro: {
+    fi: '${amount} euroa',
+    sv: '${amount}€',
+    en: '€ ${amount}',
+
+  },
+  required: {
+    fi: '(Pakollinen kenttä)',
+    sv: null,
+    en: null,
+  },
+  calculate: {
+    fi: 'Laske arvio',
+    sv: 'Beräkna uppskattningen',
+    en: 'Calculate estimate',
+  },
+  reset: {
+    fi: 'Tyhjennä tiedot',
+    sv: null,
+    en: null,
+  },
+  missing_input: {
+    fi: 'Tarkistathan vielä nämä kohdat',
+    sv: null,
+    en: null,
+  },
+  error_with_link: {
+    fi: '${error}, ${labelLink}',
+    sv: null,
+    en: null,
+  },
+  select_radio: {
+    fi: 'Valinta on pakollinen',
+    sv: null,
+    en: null,
+  },
+  enter_value: {
+    fi: 'Kenttä on pakollinen',
+    sv: null,
+    en: null,
+  },
+  must_be_number: {
+    fi: 'Kenttään on syötettävä numero',
+    sv: null,
+    en: null,
+  },
+  must_be_whole_number: {
+    fi: 'Kenttään on syötettävä kokonaisluku',
+    sv: null,
+    en: null,
+  },
+  min_or_max_out_of_bounds: {
+    fi: 'Arvon pitää olla väliltä ${min} ja ${max}',
+    sv: null,
+    en: null,
+  },
+  min_out_of_bounds: {
+    fi: 'Arvon pitää olla ${min} tai enemmän',
+    sv: null,
+    en: null,
+  },
+  max_out_of_bounds: {
+    fi: 'Arvon pitää olla ${max} tai vähemmän',
+    sv: null,
+    en: null,
+  },
+  result: {
+    fi: 'Lopputulos',
+    sv: null,
+    en: 'Result',
+  },
+};
+
+export default globalTranslations;
